fix GCC -Wall -Wextra -Wno-unused-parameter on mingw-w64
Note that MSVCRT has no %e strftime(3) specifier. Unconditionally
replace with %d.

v2:

Fix typo in network.c spotted by @lemonsqueeze

+++ patternsp.c
@@ -408,7 +408,7 @@ spatial_dict_hashstats(struct spatial_dict *dict)
 			dict->collisions, dict->repetitions,
 			(double) dict->fills * 100 / buckets,
 			dict->fills, buckets,
-			sizeof(dict->hash) / 1024 / 1024);
+			(int)(sizeof(dict->hash) / 1024 / 1024));
 }
 
 void
